deployment/requirements.txt
streamlit
numpy
tensorflow-cpu
scikit-learn
streamlit_lottie
nltk